Trade logic in Ruby for logic in YAML.
I don't like this, I think it might be worth :skull:ing Mongo's
`db:test:prepare` task and replacing it with a no op one. The logic in
the config/mongo.yml file mirrors the logic in the Ruby initializer, but
no one should rightfully expect logic to exist in the yml file.