Make contributing easier and more awesome.
This pull request adds the following:

Script Bootstrap
----------------

`script/bootstrap`, conventionally used to get new clones up and running
in a single command. This one checks for Ruby and Mongo and then,
assuming both are present, installs (if necessary) and runs bundler. The
idea is that there are only two things new contributors ever need to do.

1. git clone
2. script/bootstrap

If we really like it, we can make it smart enough to recommend ways to
*get* MongoDB and Ruby based on the OS, but I feel like that's a bit
overkill for a v1. The idea is that contribution should be as smooth as
possible so would-be helpers can start helping immediately. The script
should remain idempotent so if changes are made to the application,
existing contributors can pull, run script/bootstrap, and be ready to go
without borking their existing dev instance.

One thing this does as a side effect is specify a path for vendored gems
and binstubs, this is why the .gitignore changes were made.

.ruby-version
-------------

Specify a Ruby version using the ruby manager agnostic version of an
rvmrc.

CONTRIBUTING.md
---------------

When you create a new Pull Request or Issue, GitHub will highlight a
block above the new issue/PR title calling attention to the repository's
CONTRIBUTING.md file, should one exist. This is an expanded version of
the Helping Out section from the README.